refactor(mails): création d'un module de gestion des emails feat(mails): nouvel utilisateur, nouvelle structure B
