+++ packages/backend/src/structures/structure.module.ts
@@ -1,18 +1,24 @@
-import { forwardRef, Module } from "@nestjs/common";
+import { forwardRef, Module, HttpModule } from "@nestjs/common";
 import { ConfigService } from "../config/config.service";
+
 import { DatabaseModule } from "../database/database.module";
 import { InteractionsModule } from "../interactions/interactions.module";
 import { UsagersModule } from "../usagers/usagers.module";
 import { UsersModule } from "../users/users.module";
 import { StructuresProviders } from "./structures-providers";
 import { StructuresController } from "./structures.controller";
-import { StructuresService } from "./structures.service";
+
+import { StructuresService } from "./services/structures.service";
+
+import { MailsModule } from "../mails/mails.module";
 
 @Module({
   controllers: [StructuresController],
   exports: [StructuresService, ...StructuresProviders],
   imports: [
     DatabaseModule,
+    HttpModule,
+    forwardRef(() => MailsModule),
     forwardRef(() => UsersModule),
     forwardRef(() => UsagersModule),
     forwardRef(() => InteractionsModule),
