+++ packages/backend/src/users/services/mailjet.service.ts
@@ -21,111 +21,6 @@ export class MailJetService {
     this.domifaFromMail = this.configService.get("DOMIFA_MAILJET_FROM_EMAIL");
   }
 
-  public newStructure(structure: Structure, user: User) {
-    appLogger.warn(
-      `[MailJetService] send admin mail to "${this.domifaAdminMail}" for new structure "${structure.nom}".`
-    );
-
-    const confirmationLink =
-      this.configService.get("DOMIFA_FRONTEND_URL") +
-      "structures/confirm/" +
-      structure._id +
-      "/" +
-      structure.token;
-
-    const deleteLink =
-      this.configService.get("DOMIFA_FRONTEND_URL") +
-      "structures/delete/" +
-      structure._id +
-      "/" +
-      structure.token;
-
-    return this.mailjet
-      .post("send", { version: "v3.1" })
-      .request({
-        Messages: [
-          {
-            From: {
-              Email: this.domifaFromMail,
-              Name: "Domifa",
-            },
-            Subject: "Nouvelle structure sur Domifa : " + structure.nom,
-            TemplateID: 987764,
-            TemplateLanguage: true,
-            To: [
-              {
-                Email: this.domifaAdminMail,
-                Name: "Domifa",
-              },
-            ],
-            Variables: {
-              adresse: structure.adresse,
-              agrement: structure.agrement,
-              codePostal: structure.codePostal,
-              confirmation_link: confirmationLink,
-              departement: structure.departement,
-              email: structure.email,
-              lien_confirmation: confirmationLink,
-              lien_suppression: deleteLink,
-              phone: structure.phone,
-              responsable_fonction: structure.responsable.fonction,
-              responsable_nom: structure.responsable.nom,
-              responsable_prenom: structure.responsable.prenom,
-              structure_name: structure.nom,
-              structure_type:
-                structure.structureType === "asso"
-                  ? "Organisme agrée"
-                  : "CCAS / CIAS",
-              user_email: user.email,
-              user_nom: user.nom,
-              user_prenom: user.prenom,
-              ville: structure.ville,
-            },
-          },
-        ],
-      })
-      .then((res) => {
-        // appLogger.warn(`[MailJetService] mail response: ${JSON.stringify(res)}`);
-        return res;
-      })
-      .catch((err) => {
-        appLogger.warn(`[MailJetService] mail response error`, {
-          sentryBreadcrumb: true,
-        });
-        throw err;
-      });
-  }
-
-  public newUser(admin: User, user: User) {
-    const lienConnexion =
-      this.configService.get("DOMIFA_FRONTEND_URL") + "connexion";
-    return this.mailjet.post("send", { version: "v3.1" }).request({
-      Messages: [
-        {
-          From: {
-            Email: this.domifaFromMail,
-            Name: "Domifa",
-          },
-          Subject: "Nouvelle création de compte à valider",
-          TemplateID: 988278,
-          TemplateLanguage: true,
-          To: [
-            {
-              Email: admin.email,
-            },
-          ],
-          Variables: {
-            admin_prenom: admin.prenom,
-            lien: lienConnexion,
-            user_email: user.email,
-            user_nom: user.nom,
-            user_prenom: user.prenom,
-          },
-        },
-      ],
-    });
-  }
-
   public async newPassword(user: User): Promise<any> {
     const confirmationLink =
       this.configService.get("DOMIFA_FRONTEND_URL") +
